refactor(features/commands/behavior): 改文件名 `features/command.ts` 为 `features/behaviors.ts`
